// lib/interfaces/MessageId.ts
export enum MessageId{
    PKT_C_HEART_BEAT = 1000,
    PKT_S_HEART_BEAT = 1001,
    PKT_C_TEST_ROOM_CREATE = 1002,
    PKT_S_TEST_ROOM_CREATE = 1003,
    PKT_C_TEST_ROOM_LIST = 1004,
    PKT_S_TEST_ROOM_LIST = 1005,
    PKT_C_ROOM_JOIN = 1006,
    PKT_S_ROOM_JOIN = 1007,
    PKT_C_PLAYER_ID = 1008,
    PKT_S_PLAYER_ID = 1009,
    PKT_C_GROUP_LIST = 1010,
    PKT_S_GROUP_LIST = 1011,
    PKT_C_GROUP_JOIN = 1012,
    PKT_S_GROUP_JOIN = 1013,
    PKT_C_INITIAL_OBJECTS = 1014,
    PKT_S_INITIAL_OBJECTS = 1015,
    PKT_S_OTHER_CLIENT_JOINED = 1016,
    PKT_C_ADD_NETWORK_OBJECTS = 1017,
    PKT_S_ADD_NETWORK_OBJECTS = 1018,
    PKT_C_REMOVE_NETWORK_OBJECTS = 1019,
    PKT_S_REMOVE_NETWORK_OBJECTS = 1020,
    PKT_C_UPDATE_NETWORK_OBJECTS = 1021,
    PKT_S_UPDATE_NETWORK_OBJECTS = 1022,
    PKT_C_CHANGE_OBJECTS_OWNER = 1023,
    PKT_S_CHANGE_OBJECTS_OWNER = 1024,
    PKT_C_CHAT = 1025,
    PKT_S_CHAT = 1026,
}